requirements.txt
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
transformers>=4.33.0
torch>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
reportlab>=4.0.0
markdown>=3.5.0
streamlit>=1.28.0
python-dotenv>=1.0.0
ollama>=0.1.7
